+++ custom_node_helpers/ComfyUI_Reactor_Node.py
@@ -1,5 +1,6 @@
 from custom_node_helper import CustomNodeHelper
 
+
 class ComfyUI_Reactor_Node(CustomNodeHelper):
     facedetection_weights = {
         "retinaface_resnet50": "detection_Resnet50_Final.pth",
@@ -10,16 +11,18 @@ class ComfyUI_Reactor_Node(CustomNodeHelper):
 
     @staticmethod
     def add_weights(weights_to_download, node):
-        if "class_type" in node and node["class_type"] in [
-            "ReActorFaceSwap",
-            "ReActorLoadFaceModel",
-            "ReActorSaveFaceModel",
-        ]:
+        if node.is_type_in(
+            [
+                "ReActorFaceSwap",
+                "ReActorLoadFaceModel",
+                "ReActorSaveFaceModel",
+            ]
+        ):
             weights_to_download.append("models/buffalo_l")
             weights_to_download.append("parsing_parsenet.pth")
 
-            if "facedetection" in node["inputs"]:
-                facedetection_model = node["inputs"]["facedetection"]
+            if node.has_input("facedetection"):
+                facedetection_model = node.input("facedetection")
                 if facedetection_model in ComfyUI_Reactor_Node.facedetection_weights:
                     weights_to_download.append(
                         ComfyUI_Reactor_Node.facedetection_weights[facedetection_model]