+++ custom_node_helpers/WAS_Node_Suite.py
@@ -1,13 +1,12 @@
 from custom_node_helper import CustomNodeHelper
 
+
 class WAS_Node_Suite(CustomNodeHelper):
     @staticmethod
     def add_weights(weights_to_download, node):
-        node_class = node.get("class_type")
-        model_name = node.get("inputs", {}).get("model")
         if (
-            node_class == "CLIPSeg Model Loader"
-            and model_name == "CIDAS/clipseg-rd64-refined"
+            node.is_type("CLIPSeg Model Loader")
+            and node.input("model") == "CIDAS/clipseg-rd64-refined"
         ):
             weights_to_download.extend(["models--CIDAS--clipseg-rd64-refined"])
 
@@ -24,13 +23,9 @@ class WAS_Node_Suite(CustomNodeHelper):
             "Text Random Prompt": "Makes an HTTP request out to Lexica, which is unsupported",
             "True Random.org Number Generator": "Needs an API key which cannot be supplied",
             "Image Seamless Texture": "img2texture dependency has not been added",
-            "Image Rembg (Remove Background)": "rembg dependency has not been added because it causes custom nodes to fail",
             "MiDaS Model Loader": "WAS MiDaS nodes are not currently supported",
             "MiDaS Mask Image": "WAS MiDaS nodes are not currently supported",
             "MiDaS Depth Approximation": "WAS MiDaS nodes are not currently supported",
             "Text File History Loader": "History is not persisted",
         }
-        node_class = node.get("class_type")
-        if node_class in unsupported_nodes:
-            reason = unsupported_nodes[node_class]
-            raise ValueError(f"{node_class} node is not supported: {reason}")
+        node.raise_if_unsupported(unsupported_nodes)